move all REST API data objects into a single namespace
