+++ OutlookMatters/Mattermost/DataObjects/Thread.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace OutlookMatters.Mattermost.DataObjects
+{
+    public struct Thread
+    {
+        public string[] order;
+        public Dictionary<string, Post> posts;
+    }
+}
